Don't report failure to create data folder
This is somewhat unlikely to happen, since the content/ folder is
supposed to be writable, and this simplifies the code nicely, given
that we don't want to use `e()` (or even `$e`), which are more than
doubtful (e.g. Ajax request will never see these errors).

And even if we fail to create the data folder, users will see that in
the system check of the plugin.